perf_hooks: fix truncation of uvMetricsInfo counters
libuv reports the event loop metrics exposed through
`performance.nodeTiming.uvMetricsInfo` as `uint64_t` counters, but
they were copied into an `Int32Array` (and, before that, converted
using `v8::Integer::New()`), so they wrapped around after 2^31 on
long-running processes.

Store the metrics in a `Float64Array` instead. Values are now exact
up to `Number.MAX_SAFE_INTEGER`.

+++ typings/internalBinding/performance.d.ts
@@ -146,6 +146,6 @@ export interface PerformanceBinding {
   ): InternalPerformanceBinding.ELDHistogram;
   markBootstrapComplete(): void;
   uvMetricsInfo(): void;
-  uvMetricsBuffer: Int32Array;
+  uvMetricsBuffer: Float64Array;
   now(): number;
 }
